:sparkles: Allow users to opt-out of overdrafting
To achieve this, there's a new user attribute: can_overdraw.
This is enabled by default for all new and existing users.
If this is set to false, this user's balance can go below zero.
As a first step this throws validation errors
which somehow works but is not very user-friendly.
Also, this is not yet fully exposed via the API.