:sparkles: Add WebFinger / ActivityPub avatar provider
This patch introduces a new avatar provider based on the changes from
the previous two commits.

The WebFinger / ActivityPub provider ("webfinger" internally)
interprets the avatar field in a user either as an RFC 7565 acct URI
or, if it does not contain an "@" symbol, a RFC 3986 host string.

Based on this it will contact the host and query the WebFinger endpoint.

If the WebFinger response contains an avatar relation, the link contained
in that relation will used.

In most cases howeve, for example when querying a Mastodon, Akkoma or
Calckey user, the WebFinger response will not contain such a link.
Then, the provider will look for an ActivityPub actor using an
"application/activity+json" link in the WebFinger response.

This patch implements two strategies for obtaining an URL for the
avatar image of an ActivityPub actor:

With no additional information given, mete will try to contact the
hosting server directly. This might fail for various reasons such
as missing request authentication or signature.

Given a Mastodon client access token, mete will contact the according
Mastodon server and ask it for the profile picture of the actor.
Note that depending on the server's settings, this might cause the
profile picture to be fetched from the server's cache instead of the
actor's server directly.

The results of these lookups will be cached for one day and can be
invalidated by updating the user from the UI.